Add timer to check incoming buffers
On my device it happens that after a random time (between 10 minutes and 3 hours),
it happens that no buffers are coming in anymore from v4l2src.
Therefore add a probe on v4l2src that counts the incoming frames and use a timer
that checks this counter every 3 seconds. If no frames where received during these
3 seconds, it doesn't make sense to continue so exit the application then...

Signed-off-by: Sam Van Den Berge <sam.van.den.berge@telenet.be>